# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/engine/storagegroup/StorageGroupProcessor.java
#	server/src/main/java/org/apache/iotdb/db/tools/memestimation/MemEstTool.java
#	session/src/test/java/org/apache/iotdb/session/SessionUT.java
#	tsfile/src/main/java/org/apache/iotdb/tsfile/write/schema/MeasurementSchema.java

+++ server/src/main/java/org/apache/iotdb/db/engine/tsfilemanagement/utils/HotCompactionUtils.java
@@ -206,7 +206,7 @@ public class HotCompactionUtils {
           if (newChunkMetadata != null && newChunk != null) {
             // wait for limit write
             MergeManager.mergeRateLimiterAcquire(compactionRateLimiter,
-                newChunk.getHeader().getDataSize() + newChunk.getData().position());
+                (long)newChunk.getHeader().getDataSize() + newChunk.getData().position());
             writer.writeChunk(newChunk, newChunkMetadata);
             targetResource.updateStartTime(deviceId, newChunkMetadata.getStartTime());
             targetResource.updateEndTime(deviceId, newChunkMetadata.getEndTime());
